Fixes#8
