+++ cfme/tests/containers/test_basic_metrics.py
@@ -6,6 +6,7 @@ from utils.version import current_version
 
 pytestmark = [
     pytest.mark.uncollectif(lambda provider: current_version() < "5.6"),
+    pytest.mark.usefixtures('setup_provider'),
     pytest.mark.tier(1)]
 pytest_generate_tests = testgen.generate(
     testgen.container_providers, scope='function')
