+++ cfme/tests/containers/test_topology.py
@@ -2,15 +2,12 @@
 import pytest
 from utils import testgen
 from utils.version import current_version
-from utils.browser import browser, ensure_browser_open
 from cfme.web_ui.topology import Topology
 from cfme.containers.topology import Topology as ContainerTopology
 from cfme.fixtures.pytest_selenium import is_displayed_text
-from selenium.webdriver.common import keys
-from selenium.webdriver.common.action_chains import ActionChains
 from random import choice
-import time
 from utils.wait import wait_for
+from utils.browser import WithZoom
 
 pytestmark = [
     pytest.mark.uncollectif(
@@ -21,43 +18,10 @@ pytest_generate_tests = testgen.generate(
     testgen.container_providers, scope="function")
 
 
-class zoom(object):
-    """We need this class since sometimes items are out of the view
-    Args:
-        * level: int, the zooming value
-                (i.e. -2 -> 2 clicks out; 3 -> 3 clicks in)
-    """
-    def __init__(self, level):
-        self._level = level
-
-    def __call__(self, func):
-        def wrapper(*args, **kwargs):
-            ensure_browser_open()
-            with self:
-                return func(*args, **kwargs)
-        return wrapper
-
-    def __enter__(self, *args, **kwargs):
-        AC = ActionChains(browser())
-        for _ in xrange(abs(self._level)):
-            AC.send_keys(keys.Keys.CONTROL,
-                         keys.Keys.SUBTRACT if self._level < 0 else keys.Keys.ADD)
-        AC.perform()
-        time.sleep(1)
-
-    def __exit__(self, *args, **kwargs):
-        AC = ActionChains(browser())
-        for _ in xrange(abs(self._level)):
-            AC.send_keys(keys.Keys.CONTROL,
-                         keys.Keys.SUBTRACT if -self._level < 0 else keys.Keys.ADD)
-        AC.perform()
-        time.sleep(1)
-
-
 # CMP-9996
 
 
-@zoom(-4)
+@WithZoom(-4)
 def test_topology_display_names():
     """Testing Display Names functionality in Topology view/
 
@@ -77,7 +41,7 @@ def test_topology_display_names():
 # CMP-9998
 
 
-@zoom(-4)
+@WithZoom(-4)
 def test_topology_search():
     """Testing search functionality in Topology view.
 
@@ -105,7 +69,7 @@ def test_topology_search():
 # CMP-9999
 
 
-@zoom(-4)
+@WithZoom(-4)
 def test_topology_toggle_display():
     """Testing display functionality in Topology view.
